Fixed an integer rollover bug in the evict task
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@511 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-app/src/main/java/org/sonatype/nexus/tasks/EvictUnusedProxiedItemsTask.java
@@ -55,19 +55,19 @@ public class EvictUnusedProxiedItemsTask
         if ( getRepositoryGroupId() != null )
         {
             return getNexus().evictRepositoryGroupUnusedProxiedItems(
-                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * 24 * 60 * 60 * 1000 ),
+                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * 24 * 60 * 60 * 1000 ),
                 getRepositoryGroupId() );
         }
         else if ( getRepositoryId() != null )
         {
             return getNexus().evictRepositoryUnusedProxiedItems(
-                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * 24 * 60 * 60 * 1000 ),
+                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * 24 * 60 * 60 * 1000 ),
                 getRepositoryId() );
         }
         else
         {
             return getNexus().evictAllUnusedProxiedItems(
-                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * 24 * 60 * 60 * 1000 ) );
+                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * 24 * 60 * 60 * 1000 ) );
         }
     }
